+++ lib/src/functions/meta.dart
@@ -8,6 +8,8 @@ import 'package:collection/collection.dart';
 
 import '../ast/sass/statement/mixin_rule.dart';
 import '../callable.dart';
+import '../deprecation.dart';
+import '../evaluation_context.dart';
 import '../util/map.dart';
 import '../value.dart';
 
@@ -26,6 +28,10 @@ final global = UnmodifiableListView([
   // evaluator, because they need access to context that's only available at
   // runtime.
   _function("feature-exists", r"$feature", (arguments) {
+    warnForDeprecation(
+        "The feature-exists() function is deprecated.\n\n"
+        "More info: https://sass-lang.com/d/feature-exists",
+        Deprecation.featureExists);
     var feature = arguments[0].assertString("feature");
     return SassBoolean(_features.contains(feature.text));
   }),